Monogame base files added
My very first problem to solve is: which language, framework and/or engine should I use?
Considering the small scope of the project and my intention of displaying executed code at runtime, I put aside engines like Unity and Godot in which lots of the code being executed "under the hood" is inaccessible by external users.
However, I need to use a reflective language to be able to easily access code infos at runtime and C# still seems like a good candidate.
This led me to Monogame, a lightweight C# framework offering just the necessary basic tools to start working right away. Additionally, this is a framework I wanted to try for a while so now seems like the right time!